+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Master
+
+- Remove `pt60` class from `ExamplesPage` component. ([#70](https://github.com/mapbox/dr-ui/pull/70))
+- Update styles for `Card` and `CardContainer` components for consistent padding and spacing. ([#71](https://github.com/mapbox/dr-ui/pull/71))
+
 ## 0.0.9
 
 - Add missing `index.js` file for the `LevelIndicator` component.